[docs] Share search replay logic between search surfaces
SearchDialog and MobileNavDrawer carried identical copies of the
replay-when-ready effect, its ref bookkeeping, and the groupBy options.
Move them into useDocsSearch, which already owns isReady and search,
and expose performSearch for input changes.
